moved contract wasm & abi's back to `api-contract`
