+++ packages/api-contract/test/contracts/erc20.wasm
