+++ packages/api-contract/test/contracts/flipper-pruned.wasm
