+++ packages/api-contract/test/contracts/shared_vec-pruned.wasm
